refactor(telemetry): co-locate toMarkdown for PingPreview, OptOut, Receipt
OptOut and Receipt are server components, so toMarkdown attaches
as a static method on the same file. PingPreview is "use client";
its import becomes a client reference on the server, so static
methods aren't callable from the resolver. PingPreview's markdown
form lives in a sibling ping-preview.md.ts.
