+++ lib/telemetry-md.ts
@@ -1,13 +1,10 @@
 import { _markdown } from "@/content/telemetry.mdx";
 import { renderPlaceholder } from "fumadocs-core/mdx-plugins/remark-llms.runtime";
+
+import { OptOut } from "@/components/telemetry/opt-out";
+import { pingPreviewToMarkdown } from "@/components/telemetry/ping-preview.md";
+import { Receipt } from "@/components/telemetry/receipt";
 import { Rights } from "@/components/telemetry/rights";
-import {
-  EMISSION_GATES,
-  FIELDS,
-  NEVER_SENT,
-  OPT_OUT_PATHS,
-  RETENTION,
-} from "@/lib/telemetry-data";
 
 /**
  * Resolve the remarkLLMs markdown export from telemetry.mdx,
@@ -17,62 +14,9 @@ import {
  */
 export function resolveTelemetryMarkdown(): Promise<string> {
   return renderPlaceholder(_markdown, {
-    PingPreview() {
-      const lines = ["```", "{"];
-      for (const f of FIELDS) {
-        lines.push(`  "${f.name}": "${f.example}",`);
-      }
-      lines.push("}", "```");
-      return lines.join("\n");
-    },
-
+    PingPreview: pingPreviewToMarkdown,
     Rights: Rights.toMarkdown,
-
-    OptOut() {
-      const lines = [
-        "## How to opt out",
-        "",
-        ...OPT_OUT_PATHS.inApp.map((p) => `- **${p.label}.** ${p.description}`),
-        "",
-        `For locked-down deployments and CI images: ${OPT_OUT_PATHS.envVar}`,
-      ];
-      return lines.join("\n");
-    },
-
-    Receipt() {
-      const lines: string[] = [];
-
-      // Fields table
-      lines.push("## Exactly what's sent", "");
-      lines.push("| Field | Example | Description |", "|---|---|---|");
-      for (const f of FIELDS) {
-        lines.push(`| ${f.name} | \`${f.example}\` | ${f.description} |`);
-      }
-
-      // Never sent
-      lines.push("", "## What is never sent or stored", "");
-      for (const n of NEVER_SENT) lines.push(`- ${n}`);
-
-      // Emission gates
-      lines.push("", "## When a ping is suppressed", "");
-      lines.push("| Gate | Trigger |", "|---|---|");
-      for (const g of EMISSION_GATES) {
-        lines.push(`| ${g.name} | ${g.trigger} |`);
-      }
-
-      // Retention
-      lines.push("", "## Retention and schema evolution", "");
-      lines.push(
-        `- **Raw pings**: kept for ${RETENTION.rawPingDays} days, then deleted by a nightly cleanup job.`,
-      );
-      lines.push(
-        `- **Daily aggregate counts**: kept ${RETENTION.aggregatesKept}. No \`install_id\`; just counts grouped by ${RETENTION.rollupKeys.map((k) => "`" + k + "`").join(", ")}.`,
-      );
-      lines.push(
-        "- New fields may be added over time (additive only). Any field removal is a breaking change that gets a new route version (`/v2/ping`).",
-      );
-
-      return lines.join("\n");
-    },
+    OptOut: OptOut.toMarkdown,
+    Receipt: Receipt.toMarkdown,
   });
 }
